+++ src/js/express-xr/src/utils/xrControllerFuncs.js
@@ -50,9 +50,44 @@ const constraintObjectRotation = (controller, worldScale) => {
   }
 }
 
+const setControllerData = controller => {
+  const geometry = new THREE.BufferGeometry().setFromPoints([new THREE.Vector3(0, 0, 0), new THREE.Vector3(0, 0, -1)])
+  const material = new THREE.LineBasicMaterial({
+    color: 0x0000ff,
+    depthTest: false,
+    depthWrite: false,
+    transparent: true
+  })
+
+  const line = new THREE.Line(geometry, material)
+  line.name = 'line'
+  line.scale.z = 5
+  line.rotation.x = (Math.PI / 180) * -45
+  controller.add(line)
+
+  const raycastTiltGroup = new THREE.Group()
+  const raycastDepth = new THREE.Mesh(new THREE.BoxGeometry(0.1, 0.1, 0.1), new THREE.MeshBasicMaterial())
+  raycastDepth.visible = false
+  raycastDepth.name = 'raycast-depth'
+  raycastTiltGroup.rotation.x = (Math.PI / 180) * -45
+  raycastTiltGroup.add(raycastDepth)
+
+  controller.add(raycastTiltGroup)
+
+  controller.intersections = []
+  controller.pressed = false
+  controller.gripped = false
+  controller.interaction = {
+    grip: undefined,
+    press: undefined,
+    hover: undefined
+  }
+}
+
 module.exports = {
   getIntersections,
   boneIntersect,
   intersectObjects,
-  constraintObjectRotation
+  constraintObjectRotation,
+  setControllerData
 }
